+++ src/features/posts/ui/TagSelect.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useState } from "react"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui/select"
 
 interface TagSelectProps {
@@ -5,10 +6,26 @@ interface TagSelectProps {
   setSelectedTag: (tag: string) => void
   fetchPostsByTag: (tag: string) => void
   updateURL: () => void
-  tags: any[]
 }
 
-const TagSelect = ({ selectedTag, setSelectedTag, fetchPostsByTag, updateURL, tags }: TagSelectProps) => {
+const TagSelect = ({ selectedTag, setSelectedTag, fetchPostsByTag, updateURL }: TagSelectProps) => {
+  const [tags, setTags] = useState([])
+
+  // 태그 가져오기
+  const fetchTags = async () => {
+    try {
+      const response = await fetch("/api/posts/tags")
+      const data = await response.json()
+      setTags(data)
+    } catch (error) {
+      console.error("태그 가져오기 오류:", error)
+    }
+  }
+
+  useEffect(() => {
+    fetchTags()
+  }, [])
+
   return (
     <Select
       value={selectedTag}
