+++ src/widgets/posts/ui/PostTable.tsx
@@ -11,9 +11,9 @@ interface PostTableProps {
   updateURL: () => void
   openUserModal: (user: any) => void
   openPostDetail: (post: any) => void
-  deletePost: (id: string) => void
   setSelectedPost: (post: any) => void
   setShowEditDialog: (show: boolean) => void
+  setPosts: (posts: any[]) => void
 }
 
 const PostTable = ({
@@ -24,10 +24,21 @@ const PostTable = ({
   updateURL,
   openUserModal,
   openPostDetail,
-  deletePost,
   setSelectedPost,
   setShowEditDialog,
+  setPosts,
 }: PostTableProps) => {
+  const deletePost = async (id: number) => {
+    try {
+      await fetch(`/api/posts/${id}`, {
+        method: "DELETE",
+      })
+      setPosts(posts.filter((post) => post.id !== id))
+    } catch (error) {
+      console.error("게시물 삭제 오류:", error)
+    }
+  }
+
   return (
     <Table>
       <TableHeader>
